+++ src/components/CountrySelector/index.tsx
@@ -1,12 +1,11 @@
 import { h, Component } from 'preact'
 import { Button } from '@onfido/castor-react'
 import classNames from 'classnames'
+
+import ScreenLayout from '../Theme/ScreenLayout'
 import PageTitle from '../PageTitle'
-import { localised /* , type LocalisedType */ } from '../../locales'
-import {
-  getSupportedCountriesForDocument,
-  /* type CountryData, */
-} from '../../supported-documents'
+import { useLocales } from '~locales'
+import { getSupportedCountriesForDocument } from '../../supported-documents'
 import { trackComponent } from 'Tracker'
 import { parseTags, preventDefaultOnClick } from '~utils'
 import { hasOnePreselectedDocument } from '~utils/steps'
@@ -15,45 +14,53 @@ import Autocomplete from 'accessible-autocomplete/preact'
 import theme from 'components/Theme/style.scss'
 import style from './style.scss'
 
-/* type Props = {
-  documentType: string,
-  idDocumentIssuingCountry: CountryData,
-  previousStep: () => void,
-  nextStep: () => void,
-  translate: (string, ?{}) => string,
-  actions: Object,
-} & LocalisedType
+import type { CountryData } from '~types/commons'
+import type { WithLocalisedProps, WithTrackingProps } from '~types/hocs'
+import type { StepComponentBaseProps } from '~types/routers'
+
+export type Props = {
+  documentType: string
+  idDocumentIssuingCountry: CountryData
+  previousStep: () => void
+  nextStep: () => void
+} & WithLocalisedProps &
+  WithTrackingProps &
+  StepComponentBaseProps
 
 type State = {
-  showNoResultsError: Boolean,
-} */
+  showNoResultsError: boolean
+}
 
-const getFlagIconURL = (country) => {
+const getFlagIconURL = (country: CountryData) => {
   // NOTE: `flagsPath` is the same as what is returned by libphonenumber-js in PhoneNumberInput component
   const flagsPath = 'https://lipis.github.io/flag-icon-css/flags/4x3/'
   return `${flagsPath}${country.country_alpha2.toLowerCase()}.svg`
 }
 
-const getCountryOptionTemplate = (country) =>
-  `<i
+const getCountryOptionTemplate = (country: CountryData) => {
+  if (country) {
+    return `<i
       role="presentation"
       class="${style.countryFlag}"
       style="background-image: url(${getFlagIconURL(country)})"></i>
-    <span class="${style.countryLabel}">${country.name}</span>`
+      <span class="${style.countryLabel}">${country.name}</span>`
+  }
+  return ''
+}
 
-class CountrySelection extends Component {
+class CountrySelection extends Component<Props, State> {
   state = {
     showNoResultsError: false,
   }
 
-  handleCountrySearchConfirm = (selectedCountry) => {
+  handleCountrySearchConfirm = (selectedCountry: CountryData) => {
     const { actions, idDocumentIssuingCountry } = this.props
     if (selectedCountry) {
       this.setState({
         showNoResultsError: false,
       })
       actions.setIdDocumentIssuingCountry(selectedCountry)
-      setTimeout(() => document.getElementById('country-search').blur(), 0)
+      setTimeout(() => document.getElementById('country-search')?.blur(), 0)
     } else if (
       !selectedCountry &&
       (!idDocumentIssuingCountry || !idDocumentIssuingCountry.country_alpha3)
@@ -64,7 +71,10 @@ class CountrySelection extends Component {
     }
   }
 
-  suggestCountries = (query = '', populateResults) => {
+  suggestCountries = (
+    query = '',
+    populateResults: (results: CountryData[]) => string[]
+  ) => {
     const { documentType, idDocumentIssuingCountry, actions } = this.props
     if (idDocumentIssuingCountry && query !== idDocumentIssuingCountry.name) {
       actions.resetIdDocumentIssuingCountry()
@@ -78,13 +88,28 @@ class CountrySelection extends Component {
     populateResults(filteredResults)
   }
 
+  handleMenuMouseClick = (event: Event) => {
+    const target = event.target as HTMLUListElement
+    // Intercept mouse click if event target is the displayed menu, i.e. scrollbar area
+    // (mouse clicks in on the menu list options the target will be a different class)
+    // Otherwise accessible-autocomplete picks up a mouse click on scrollbar area as a confirm event
+    if (
+      target.className.includes(
+        'onfido-sdk-ui-CountrySelector-custom__menu--visible'
+      )
+    ) {
+      event.preventDefault()
+    }
+  }
+
   componentDidMount() {
     if (this.props.idDocumentIssuingCountry) {
       this.props.actions.resetIdDocumentIssuingCountry()
     }
+    document.addEventListener('mousedown', this.handleMenuMouseClick)
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps: Props) {
     if (
       prevProps.documentType &&
       this.props.documentType !== prevProps.documentType
@@ -93,16 +118,22 @@ class CountrySelection extends Component {
     }
   }
 
+  componentWillUnmount() {
+    document.removeEventListener('mousedown', this.handleMenuMouseClick)
+  }
+
   isDocumentPreselected() {
     const { steps, documentType } = this.props
     return hasOnePreselectedDocument(steps) && documentType !== 'passport'
   }
 
-  getNoResultsTextForDropdown = () =>
-    parseTags(
-      this.props.translate('country_select.alert_dropdown.country_not_found'),
+  getNoResultsTextForDropdown = () => {
+    const { translate } = useLocales()
+    return parseTags(
+      translate('country_select.alert_dropdown.country_not_found'),
       ({ text }) => text
     )
+  }
 
   trackChooseAnotherDocumentTypeClick = () => {
     const { trackScreen, previousStep } = this.props
@@ -111,9 +142,8 @@ class CountrySelection extends Component {
   }
 
   renderNoResultsError = () => {
-    const noResultsErrorCopy = this.props.translate(
-      'country_select.alert.another_doc'
-    )
+    const { translate } = useLocales()
+    const noResultsErrorCopy = translate('country_select.alert.another_doc')
 
     return (
       <div className={style.errorContainer}>
@@ -136,9 +166,27 @@ class CountrySelection extends Component {
   }
 
   render() {
-    const { translate, nextStep, idDocumentIssuingCountry } = this.props
+    const { translate } = useLocales()
+    const { nextStep, idDocumentIssuingCountry } = this.props
+
     return (
-      <div className={theme.fullHeightContainer}>
+      <ScreenLayout
+        actions={
+          <Button
+            variant="primary"
+            className={classNames(theme['button-centered'], theme['button-lg'])}
+            disabled={
+              !idDocumentIssuingCountry ||
+              !idDocumentIssuingCountry.country_alpha3 ||
+              this.state.showNoResultsError
+            }
+            onClick={nextStep}
+            data-onfido-qa="countrySelectorNextStep"
+          >
+            {translate('country_select.button_primary')}
+          </Button>
+        }
+      >
         <PageTitle title={translate('country_select.title')} />
         <div className={classNames(theme.alignTextLeft, style.container)}>
           <div data-onfido-qa="countrySelector">
@@ -155,9 +203,9 @@ class CountrySelection extends Component {
               displayMenu="overlay"
               cssNamespace={'onfido-sdk-ui-CountrySelector-custom'}
               templates={{
-                inputValue: (country) => country && country.name,
-                suggestion: (country) =>
-                  country && getCountryOptionTemplate(country),
+                inputValue: (country: CountryData) => country?.name,
+                suggestion: (country: CountryData) =>
+                  getCountryOptionTemplate(country),
               }}
               onConfirm={this.handleCountrySearchConfirm}
             />
@@ -166,24 +214,9 @@ class CountrySelection extends Component {
             this.state.showNoResultsError &&
             this.renderNoResultsError()}
         </div>
-        <div>
-          <Button
-            variant="primary"
-            className={classNames(theme['button-centered'], theme['button-lg'])}
-            disabled={
-              !idDocumentIssuingCountry ||
-              !idDocumentIssuingCountry.country_alpha3 ||
-              this.state.showNoResultsError
-            }
-            onClick={nextStep}
-            data-onfido-qa="countrySelectorNextStep"
-          >
-            {translate('country_select.button_primary')}
-          </Button>
-        </div>
-      </div>
+      </ScreenLayout>
     )
   }
 }
 
-export default trackComponent(localised(CountrySelection), 'country_select')
+export default trackComponent(CountrySelection, 'country_select')