+++ src/components/Router/StepComponentMap.ts
@@ -243,7 +243,8 @@ const buildNonPassportPreCaptureComponents = (
     ? []
     : [SelectIdentityDocument]
   const prependCountrySelector = showCountrySelection ? [CountrySelector] : []
-
+  // @ts-ignore
+  // TODO: convert DocumentSelector to TS
   return [...prependDocumentSelector, ...prependCountrySelector]
 }